+++ Violin-black.pngZone.Identifier
@@ -1,0 +1,4 @@
+[ZoneTransfer]
+ZoneId=3
+ReferrerUrl=file:///C:/Users/gabri/AppData/Local/Programs/Microsoft%20VS%20Code/resources/app/out/vs/code/electron-browser/workbench/workbench.html
+HostUrl=file://wsl%24/Ubuntu-18.04/home/gabrielle_tobermann/workspace/projects/personal/Personal-bio-site-webpack/Violin-black.png
