+++ src/components/OrderForm.js
@@ -71,9 +71,8 @@ function OrderForm({
     createOrderAndOrderItems(order.transactionID, itemInputs, order).then((resp) => {
       setOrders(resp.order);
       setOrderItems(resp.orderItems);
-      console.warn('orders', resp.order);
-      console.warn('orderItems', resp.orderItems);
-      console.warn(resp);
+      console.warn('resp.order', resp.order);
+      console.warn('resp.orderItems', resp.orderItems);
     });
   };
